+++ src/validation.h
@@ -558,7 +558,7 @@ extern CBlockTreeDB *pblocktree;
 extern tp::ThreadPool* threadpool;
 extern std::vector<std::pair<uint256, int64_t> > vecTPSTestReceivedTimesMempool;
 extern int64_t nTPSTestingStartTime;
-extern int64_t nTPSTestingSendRawEndTime;
+extern double nTPSTestingSendRawEndTime;
 extern int64_t nTPSTestingSendRawStartTime;
 extern std::vector<JSONRPCRequest> vecTPSRawTransactions;
 /**
